SortingRequest and FilterRequest moved to service layer
