+++ service/src/main/java/com/epam/rft/atsy/service/CandidateService.java
@@ -1,6 +1,6 @@
 package com.epam.rft.atsy.service;
 
-import com.epam.rft.atsy.persistence.request.FilterRequest;
+import com.epam.rft.atsy.service.request.FilterRequest;
 import com.epam.rft.atsy.service.domain.CandidateDTO;
 
 import java.util.Collection;
