+++ service/src/main/java/com/epam/rft/atsy/service/request/SearchOptions.java
@@ -1,4 +1,4 @@
-package com.epam.rft.atsy.persistence.request;
+package com.epam.rft.atsy.service.request;
 
 import com.google.common.base.MoreObjects;
 import org.apache.commons.lang3.StringUtils;
